refactor: standardize block context usage and improve method signatures

+++ BetaSharp/Blocks/BlockSoulSand.cs
@@ -7,7 +7,6 @@ namespace BetaSharp.Blocks;
 
 internal class BlockSoulSand : Block
 {
-
     public BlockSoulSand(int id, int textureId) : base(id, textureId, Material.Sand)
     {
     }
@@ -15,7 +14,7 @@ internal class BlockSoulSand : Block
     public override Box? getCollisionShape(IBlockReader world, int x, int y, int z)
     {
         float height = 2.0F / 16.0F;
-        return new Box((double)x, (double)y, (double)z, (double)(x + 1), (double)((float)(y + 1) - height), (double)(z + 1));
+        return new Box(x, y, z, x + 1, y + 1 - height, z + 1);
     }
 
     public override void onEntityCollision(World world, int x, int y, int z, Entity entity)
